Bump versions, put plugin and services functions into singletons
