+++ gradle.properties
@@ -1,5 +1,4 @@
 group=com.mineinabyss
-version=0.13
-kotlinVersion=1.6.10
+version=0.14
+kotlinVersion=1.7.10
 serverVersion=1.19.2-R0.1-SNAPSHOT
-copyJar=false
