+++ gradle/libs.versions.toml
@@ -1,26 +1,26 @@
 [versions]
 anvilgui = "1.5.3-SNAPSHOT"
-coroutines = "1.6.1"
-exposed = "0.38.2"
+coroutines = "1.6.4"
+exposed = "0.39.2"
 fastutil = "8.2.2"
-fawe = "2.4.6"
-junit = "5.8.2"
-kaml = "0.43.0"
-kmongo = "4.3.0"
-koin = "3.1.6"
-kotest = "5.2.3"
-kotlin = "1.6.10"
+fawe = "2.4.5"
+junit = "5.9.0"
+kaml = "0.47.0"
+kmongo = "4.7.1"
+koin = "3.2.1"
+kotest = "5.4.2"
+kotlin = "1.7.10"
 kotlin-statistics = "1.2.1"
-ktor = "2.0.1"
-logback = "1.2.3"
+ktor = "2.1.1"
+logback = "1.4.1"
 mccoroutine = "2.5.0"
-mockbukkit = "1.24.1"
-mockk = "1.12.3"
+mockbukkit = "2.120.1"
+mockk = "1.13.1"
 modelengine = "R3.0.0"
 protocollib = "5.0.0-SNAPSHOT"
 reflections = "0.9.12"
-serialization = "1.3.2"
-sqlite-jdbc = "3.36.0.2"
+serialization = "1.4.0"
+sqlite-jdbc = "3.39.3.0"
 vault = "1.7"
 worldguard = "7.1.0-SNAPSHOT"
 
@@ -30,7 +30,7 @@ minecraft-headlib-core = "de.erethon:headlib:3.0.10"
 minecraft-mccoroutine = { module = "com.github.shynixn.mccoroutine:mccoroutine-bukkit-api", version.ref = "mccoroutine" }
 minecraft-mccoroutine-core = { module = "com.github.shynixn.mccoroutine:mccoroutine-bukkit-core", version.ref = "mccoroutine" }
 minecraft-anvilgui = { module = "net.wesjd:anvilgui", version.ref = "anvilgui" }
-minecraft-mockbukkit = { module = "com.github.seeseemelk:MockBukkit-v1.18", version.ref = "mockbukkit" }
+minecraft-mockbukkit = { module = "com.github.seeseemelk:MockBukkit-v1.19", version.ref = "mockbukkit" }
 
 minecraft-plugin-modelengine = { module = "com.ticxo.modelengine:api", version.ref = "modelengine" }
 minecraft-plugin-protocollib = { module = "com.comphenix.protocol:ProtocolLib", version.ref = "protocollib" }
@@ -89,8 +89,7 @@ kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref 
 kotest-property = { module = "io.kotest:kotest-property", version.ref = "kotest" }
 
 junit-bom = { module = "org.junit:junit-bom", version.ref = "junit" }
-
-junit-jupiter = { module = "org.junit.jupiter:junit-jupiter" }
+junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version.ref = "junit" }
 
 [bundles]
 minecraft = ["minecraft-headlib", "minecraft-mccoroutine", "minecraft-anvilgui"]
