+++ idofront-util/build.gradle.kts
@@ -7,4 +7,5 @@ plugins {
 
 dependencies {
     compileOnly(libs.kotlinx.serialization.json)
+    compileOnly(libs.koin.core)
 }
