+++ idofront-util/src/main/kotlin/com/mineinabyss/idofront/plugin/Registration.kt
@@ -1,6 +1,5 @@
 package com.mineinabyss.idofront.plugin
 
-import org.bukkit.Bukkit
 import org.bukkit.event.Listener
 import org.bukkit.plugin.Plugin
 import org.bukkit.plugin.PluginManager
@@ -13,7 +12,7 @@ import org.bukkit.plugin.ServicesManager
  *
  * @see ServicesManager.register
  */
-inline fun <reified T : Any> Plugin.registerService(impl: T, priority: ServicePriority = ServicePriority.Lowest) =
+inline fun <reified T : Any> Plugin.service(impl: T, priority: ServicePriority = ServicePriority.Lowest) =
     server.servicesManager.register(T::class.java, impl, this, priority)
 
 /**
@@ -21,57 +20,5 @@ inline fun <reified T : Any> Plugin.registerService(impl: T, priority: ServicePr
  *
  * @see PluginManager.registerEvents
  */
-fun Plugin.registerEvents(vararg listeners: Listener) =
+fun Plugin.listeners(vararg listeners: Listener) =
     listeners.forEach { server.pluginManager.registerEvents(it, this) }
-
-/**
- * Gets a service registered with Bukkit of type [T], throws an error if not found.
- *
- * @see ServicesManager.load
- * @see getServiceOrNull
- */
-inline fun <reified T> getService(): T =
-    getServiceOrNull<T>(null) ?: error("Could not load service ${T::class.simpleName}")
-
-/**
- * Gets a service registered with Bukkit of type [T] or null if not found.
- *
- * @param plugin The name of the plugin from which to get this service. This is used to check whether that plugin is
- * enabled before trying to load the service, thus preventing any exceptions.
- * @see ServicesManager.load
- */
-inline fun <reified T> getServiceOrNull(plugin: String? = null): T? =
-    if (plugin == null || isPluginEnabled(plugin))
-        Bukkit.getServer().servicesManager.load(T::class.java)
-    else null
-
-/**
- * Gets a service in a way that will ignore class differences across different classloaders.
- *
- * The returned type may then be cast to some common class that is shared between both loaders, such
- * as a Function type in Kotlin.
- */
-inline fun <reified T> getServiceViaClassNameOrNull(): Any? {
-    val serviceManager = Bukkit.getServer().servicesManager
-    val className = T::class.java.name
-    val clazz = serviceManager.knownServices.first { it.name == className }
-    return serviceManager.load(clazz)
-}
-
-
-/**
- * Gets a plugin registered with Bukkit of type [T], throws an error if not found.
- * @see getPluginOrNull
- */
-inline fun <reified T : Plugin> getPlugin(): T =
-    getPluginOrNull<T>(T::class.simpleName!!) ?: error("Could not find plugin ${T::class.simpleName}")
-
-/** Gets a plugin registered with Bukkit of type [T] or null if not found. */
-inline fun <reified T : Plugin> getPluginOrNull(plugin: String): T? =
-    if (isPluginEnabled(plugin))
-        Bukkit.getPluginManager().getPlugin(plugin) as? T
-    else null
-
-inline fun isPluginEnabled(plugin: String) = Bukkit.getServer().pluginManager.isPluginEnabled(plugin)
-
-inline fun doesPluginExist(plugin: String) = Bukkit.getServer().pluginManager.getPlugin(plugin) != null
